% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Points3d.R
\name{Avp3d}
\alias{Avp3d}
\title{3d added-variable plots}
\usage{
Avp3d(
  model,
  coef1,
  coef2,
  categorical = NULL,
  data = model.frame(model),
  fit = TRUE,
  ...
)
}
\arguments{
\item{model}{an \code{lm} model}

\item{coef1}{a string that is the name of a numeric variable in the model (required)}

\item{coef2}{another numeric variable in the model (required)}

\item{categorical}{an option categorical variable}

\item{data}{default: \code{model.frame(model)}}

\item{fit}{whether to show planar fit in added-variable plot (default: TRUE)}

\item{...}{arguments passes to \code{\link{Plot3d}}}
}
\description{
3d added-variable plots
}
\examples{
library(p3d)
Init3d(cex=1.2)
fit <- lm(mpg ~ cyl + gear + wt + hp, mtcars)
Avp3d(fit, 'gear','cyl')
Avp3d(fit, 'gear','cyl','gear', fit = FALSE)
Id3d()
}
