% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin.R
\name{bend}
\alias{bend}
\title{Trimming with a spline}
\usage{
bend(x, linear = 0.8)
}
\arguments{
\item{x}{variable to have ends bent}

\item{linear}{proportion in middle left unbent, default: 0.8}
}
\description{
Pull in the ends of a variable
}
