% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{cell}
\alias{cell}
\alias{cell.default}
\title{Confidence ellipse for linear model}
\usage{
cell(object, ...)

\method{cell}{default}(
  object,
  which.coef,
  levels = 0.95,
  Scheffe = FALSE,
  dfn = 2,
  center.pch = 19,
  center.cex = 1.5,
  segments = 51,
  xlab,
  ylab,
  las = par("las"),
  col = palette()[2],
  lwd = 2,
  lty = 1,
  radius = sqrt(dfn * qf(levels, dfn, dfd)),
  add = FALSE,
  ...
)
}
\arguments{
\item{object}{object}

\item{...}{Other parameters.

Draw confidence intervals}

\item{which.coef}{coefs to be plotted.}

\item{Scheffe}{Default: FALSE}

\item{dfn}{Default: 2.}

\item{center.pch}{Default: 19.}

\item{center.cex}{Default: 1.5.}

\item{segments}{Default: 51.}

\item{xlab, }{ylab}

\item{las}{Default: \code{par(las)}.}

\item{col}{Default: \code{palette()[2]}}

\item{lwd, }{lty}

\item{radius}{Default: \code{sqrt(dfn * qf(levels, dfn, dfd))}}

\item{add}{Default: FALSE.}

\item{level}{Default: 0.95}
}
\description{
See help for car::confidence.ellipse.lm
except that 'cell' returns the points to form the ellipse
which must be plotted with plot(...,type='l') or lines(...)
Use dfn to determine Sheffe dimension, i.e. dfn = 1 to generate ordinary CIs, dfn = 2 for 2-dim CE, etc.
}
\details{
TODO: extend to 3 dimensions if which.coef has length 3
}
