% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misscode.R
\name{misscode}
\alias{misscode}
\title{Recode missing data}
\usage{
misscode(x, ...)
}
\arguments{
\item{x}{vector with possible missing values}

\item{offset}{(default .1) how far below the minimum value as a proportion of the range}
}
\description{
Recodes missing data to a value less than the minimum value -- or to a non-missing
code to indicate missingness for a factor -- to facilitate plotting in 3d
}
