% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3d-package.R
\docType{package}
\name{p3d-package}
\alias{p3d}
\alias{p3d-package}
\title{p3d: Interactive 3D Visualization of p-Value Surfaces}
\description{
Provides interactive three-dimensional visualization of p-value surfaces for single-sample t-tests. Enables simultaneous exploration of effect size, sample size, and significance level, making the continuous nature of p-values and the moderating effect of sample size visually apparent. Includes both interactive (rgl, plotly) and static (persp) rendering options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahmetkutuk1/p3d}
  \item Report bugs at \url{https://github.com/ahmetkutuk1/p3d/issues}
}

}
\author{
\strong{Maintainer}: Ahmet Kutuk \email{ahmetkutuk@gmail.com}

Authors:
\itemize{
  \item Ahmet Kutuk \email{ahmetkutuk@gmail.com}
}

}
\keyword{internal}
