% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot3d.R
\name{Plot3d}
\alias{Plot3d}
\alias{Plot3d.lm}
\alias{Plot3d.formula}
\alias{Plot3d.default}
\title{Plot points in 3D}
\usage{
Plot3d(object, ...)

\method{Plot3d}{lm}(fit, ...)

\method{Plot3d}{formula}(
  formula = attr(data, "formula"),
  data = sys.parent(),
  groups = NULL,
  subset = NULL,
  xlab = names(dd)[2],
  ylab = names(dd)[1],
  zlab = names(dd)[3],
  verbose = 0,
  col = c("blue", "#008800", "orange", "magenta", "darkcyan", "red", "darkgray",
    "darkblue", "cyan", "darkorange", "darkmagenta", "cyan", "darkred", "yellowgreen",
    "purple", "gray", "#cc4466", "slategray", "coral", "purple", "pink", "#555555",
    "whitesmoke"),
  surface = FALSE,
  fit = "smooth",
  surface.col = col,
  lines.col = "gray",
  lines.lwd = 1,
  theta = 0,
  phi = 15,
  zoom = 1,
  fov = 15,
  keep.view = FALSE,
  ...
)

\method{Plot3d}{default}(x, y, z, xlab, ylab, zlab, groups = NULL, ...)
}
\arguments{
\item{object}{To dispatch.}

\item{...}{other arguments.}

\item{fit}{Default: 'smooth'.}

\item{formula}{A formula of the form \code{y ~ x + z} or
\code{y ~ x + z | w} where \code{y}, \code{x}, \code{z}
are numeric variables and
\code{w} is a categorical variable.}

\item{data}{A data frame in which \code{formula} is evaluated.}

\item{groups}{The name of a variable to serve as an alternative
to specifying a \code{w} variable in \code{formula}.}

\item{xlab, }{ylab, zlab, groups Graphical parameters.}

\item{verbose}{Default: FALSE.}

\item{col}{Colors for levels of the \code{w} variable.}

\item{surface}{Whether to plot a surface. Default: FALSE.}

\item{surface.col}{Color for surface.}

\item{lines.col}{Default: 'gray'.}

\item{lines.lwd}{Default: 1.}

\item{theta, }{phi, zoom, fov Arguments for \code{view3d}
with defaults 0, 15, 1, 15, respectively.}

\item{keep.view}{Do not change current view. Default: FALSE.}

\item{x}{z-axis (depth) variable}

\item{y}{y-axis (vertical) variable}
}
\description{
Generic function to plot points in 3d
}
\examples{
library(p3d)
library(spida2)
library(latticeExtra)
data(Smoking)
head(Smoking)
rownames(Smoking) <- Smoking$Country

dd <- na.omit(
  subset(Smoking,
         select = c(LE, CigCon, HealthExpPC,
                    Continent))
  )
Init3d(family = 'serif', cex = 2)
Plot3d( LE ~ CigCon + HealthExpPC | Continent, dd)
Axes3d()
Id3d(pad = 1)
Id3d(pad = 1, cex = 2, col = 'black')
fitlin <- lm(LE ~ CigCon, dd)
Fit3d(fitlin)
fitlin2 <- lm(LE ~ CigCon + HealthExpPC, dd)
Fit3d(fitlin2, col = 'green')
fit <- lm(LE ~ CigCon + log(HealthExpPC) +I(CigCon^2) +
    I(log(HealthExpPC)^2) + I(CigCon*log(HealthExpPC)),
    dd)
Fit3d(fit , col = 'red')
fitl <- lm(LE ~ CigCon + HealthExpPC, dd)
# HEpCap is highly 'skewed': dense on left, long tail on right
require(lattice)
densityplot(dd$HealthExpPC)

# Useful to use a transformation to make spread more even
#  e.g. log
# First make sure all values are positive:

sort(dd$HealthExpPC)

# Do log transformation:

dd$LogHE <- log(dd$HealthExpPC)    # create log HE

densityplot( dd$LogHE )

# Also useful to have categories:

dd$HECat <- cut(dd$LogHE, 7)       # create categories
summary(Smoking)

Plot3d(LE ~ CigCon + LogHE |HECat, dd)  # condition on level of HEpC
Axes3d()
Ell3d()
Id3d(pad=1)

# Simple regression

fit.lin <- lm( LE ~ CigCon, Smoking)
Fit3d( fit.lin )
fit.lin2 <- lm(LE ~ CigCon, subset(Smoking, Continent != "Africa"))
Fit3d( fit.lin2)
Pop3d(4)

# Use multiple regression (advanced version with quadratic surface)

fit = lm( LE ~ CigCon + I(LogHE^2) +I(CigCon^2) + I(LogHE^2) + I(CigCon*LogHE), Smoking)
Fit3d( fit, col = 'red' )
Pop3d(2)

# refit omitting Africa:

fit.na = lm( LE ~ CigCon + I(LogHE^2) +I(CigCon^2) + I(LogHE^2)
             + I(CigCon*LogHE), Smoking, subset = Continent != "Africa")
Fit3d( fit.na, col = 'red' )

# Marginal relationship

Pop3d(4)    # pop twice for each fit
fit.quad <- lm( LE ~ CigCon + I(CigCon^2) , Smoking)
Fit3d( fit.quad, col = 'green')

#  A quadratic surface within each Continent (overfitting?!)

fit2 = lm( LE ~ Continent*(CigCon + HealthExpPC +I(CigCon^2) +
                             I(HealthExpPC^2) + I(CigCon*HealthExpPC)), Smoking)
Fit3d( fit2 )

#  For a 2D view of the y versus x

view3d(0,0,0)   # note the first parameter, theta, rotates clockwise in the vertical axis
# from the initial position given by view3d(0,0,0); the second
# parameter, phi, tilts the graph around a horizontal screen axis
# towards the viewer.  The third parameter, fov, affects the
# amount of 'perspective'. fov = 0 yields an orthographic
# projection
}
