% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot3d.R
\name{Plot3d_par}
\alias{Plot3d_par}
\title{Maintain a list of parameters for Plot3d}
\usage{
Plot3d_par(..., new = FALSE)
}
\arguments{
\item{...}{Parameters.}

\item{new}{Open new window if TRUE. Default: FALSE}
}
\description{
Maintain a list of parameters for Plot3d
}
