% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin.R
\name{spin}
\alias{spin}
\title{Change viewpoint}
\usage{
spin(
  theta = 0,
  phi = 15,
  fov = par3d("FOV"),
  zoom = par3d("zoom"),
  scale = par3d("scale"),
  stay = FALSE
)
}
\arguments{
\item{theta}{rotation about vertical (y) axis in degrees, default: 0}

\item{phi}{angular elevation above horizontal plane, default: 15}

\item{fov}{perspective}

\item{zoom}{magnification}

\item{scale}{changes bounding box and axes but not data}
}
\description{
Change viewpoint
}
