\name{p3d-package}
\Rdversion{1.1}
\alias{p3d-package}
\alias{p3d}
\docType{package}
\title{
Visualization of models in 3D
}
\description{
    \code{p3d} is a very rough collection of functions to help visualize models
    that are expressed as functions of two independent variables with the
    addition of a possible categorical variable.
}
\details{
\tabular{ll}{
Package: \tab p3d\cr
Type: \tab Package\cr
Version: \tab 0.05\cr
Date: \tab 2009-05-09\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Georges Monette

Maintainer: Georges Monette <georges@yorku.ca>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
\section{Changes}{
\item{Lines3d.default}{2009-05-09: disabled line changing 'lwd' to 'size' since 'segments3d' uses 'lwd'.}
}

